change discount amount to double

+++ services/Discount/Discount.Application/Commands/CreateDiscountCommand.cs
@@ -12,6 +12,6 @@ namespace Discount.Application.Commands
     {
         public string ProductName { get; set; }
         public string Description { get; set; }
-        public int DiscountAmount { get; set; }
+        public double DiscountAmount { get; set; }
     }
 }
